Consolidate CLI scenario test helpers
