+++ internal/cli/conclusion_show_test.go
@@ -101,20 +101,10 @@ var _ = Describe("conclusion show JSON", func() {
 
 		got := runCLIJSON[conclusionShowJSON](dir, "conclusion", "show", "C-0001")
 		Expect(got.Conclusion).NotTo(BeNil())
-		Expect(got.ID).To(Equal("C-0001"))
-		Expect(got.CandidateRef).To(Equal("refs/heads/candidate/E-0001-a1"))
-		Expect(got.CandidateSHA).To(Equal("0123456789abcdef0123456789abcdef01234567"))
-		Expect(got.CandidateAttempt).To(Equal(2))
-		Expect(got.BaselineAttempt).To(Equal(1))
-		Expect(got.BaselineRef).To(Equal("refs/heads/baseline/E-0000-a1"))
-		Expect(got.BaselineSHA).To(Equal("89abcdef0123456789abcdef0123456789abcdef"))
-		Expect(got.Observations).To(HaveLen(4))
-		Expect(got.ObservationArtifacts).To(HaveKey("O-0001"))
-		Expect(got.ObservationArtifacts["O-0001"]).To(HaveLen(1))
-		Expect(got.ObservationArtifacts).To(HaveKey("O-0002"))
-		Expect(got.ObservationArtifacts["O-0002"]).To(BeEmpty())
-		Expect(got.ObservationArtifacts).To(HaveKey("O-0003"))
-		Expect(got.ObservationArtifacts["O-0003"]).To(HaveLen(1))
+		Expect(got.Observations).To(Equal([]string{"O-0001", "O-0002", "O-0003", "O-9999"}))
+		Expect(got.ObservationArtifacts).To(HaveKeyWithValue("O-0001", ConsistOf(HaveField("Name", "scalar"))))
+		Expect(got.ObservationArtifacts).To(HaveKeyWithValue("O-0002", BeEmpty()))
+		Expect(got.ObservationArtifacts).To(HaveKeyWithValue("O-0003", ConsistOf(HaveField("Name", "evidence/mechanism"))))
 		Expect(got.ObservationEvidenceFailures["O-0003"]).To(HaveLen(1))
 		Expect(got.ObservationEvidenceFailures["O-0003"][0].Name).To(Equal("profile"))
 		Expect(got.ObservationEvidenceFailures["O-0003"][0].ExitCode).To(Equal(7))
